requirements.txt
# Core
typer>=0.9
playwright>=1.40
pandas>=2.1
openpyxl>=3.1
pydantic>=2.5
aiohttp>=3.9
python-dotenv>=1.0
anthropic>=0.39

# Web UI
fastapi>=0.109
uvicorn[standard]>=0.27
python-multipart>=0.0.6

# Dev / testing
pytest>=7.0
pytest-asyncio>=0.23
